requirements/base.txt
aniso8601
annotated-types
asgiref
attrs
autocommand
backports.tarfile
black
certifi
charset-normalizer
click
coverage
Django
django-cors-headers
django-filter
djangorestframework
drf-spectacular
exceptiongroup
factory-boy
Faker
flake8
graphene
graphene-django
graphql-core
graphql-relay
gunicorn
idna
importlib-metadata
inflect
inflection
iniconfig
jaraco.collections
jaraco.context
jaraco.functools
jaraco.text
jsonschema
jsonschema-specifications
mccabe
more-itertools
mypy-extensions
packaging
pathspec
pillow
pip
platformdirs
pluggy
promise
psycopg2-binary
pycodestyle
pydantic
pydantic-core
pyflakes
PyJWT
pytest
pytest-cov
pytest-django
pytest-factoryboy
pytest-mock
python-dateutil
python-environ
pytz
PyYAML
qrcode
referencing
requests
rpds-py
setuptools
six
sqlparse
text-unidecode
tomli
typeguard
typing-extensions
uritemplate
urllib3
wheel
whitenoise
zipp
